Add a clearer error message for missing HuggingFace model
I got stuck on this for an hour because the error message makes it sound like the user is supposed to sign the Hugging Face EULA rather than the EULAs for the specific models. This more specific error message will make it more obvious to users which EULA they need to accept.